autogen-orchestrator/requirements.txt
pyautogen
fastapi
uvicorn
python-dotenv
requests
pydantic
selenium
webdriver-manager